[Sidebar] Migrate Sidebar.
