+++ frontend/src/components/polls/PollsBase.tsx
@@ -5,22 +5,17 @@ import { PollsSidebar } from "./PollsSidebar";
 import Container from "react-bootstrap/Container";
 
 type Props = {
-    title: any;
     children: any;
 };
 
 export function PollsBase(props: Props) {
     return (
-        <Container className="mt-3">
+        <Container className="mt-5">
             <Row>
-                <Col xs={{ offset: 2 }}>{props.title}</Col>
-            </Row>
-
-            <Row>
-                <Col xs={2}>
-                    <PollsSidebar />
+                <Col md="3">
+                    <PollsSidebar adminVersion />
                 </Col>
-                <Col>{props.children}</Col>
+                <Col md="9">{props.children}</Col>
             </Row>
         </Container>
     );
