--- frontend/src/components/sidebar.css
@@ -1,8 +1,0 @@
-.sidebar {
-    text-align: left;
-}
-
-.sidebar .btn {
-    width: 100%;
-    font-weight: 500;
-}
